refactor: move security config and services to dedicated package
