+++ back/src/main/java/com/openclassrooms/mdd/security/service/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package com.openclassrooms.mdd.service;
+package com.openclassrooms.mdd.security.service;
 
 
 import com.openclassrooms.mdd.model.User;
@@ -7,9 +7,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
-import java.util.Optional;
 
 /**
  * @author Wilhelm Zwertvaegher
@@ -24,13 +21,10 @@ public class CustomUserDetailsService implements UserDetailsService {
         this.userRepository = userRepository;
     }
 
-    // @Override
+    @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<User> foundUser = userRepository.findByEmail(username);
-        if(foundUser.isEmpty()) {
-            throw new UsernameNotFoundException(username);
-        }
-        User user = foundUser.get();
-        return new org.springframework.security.core.userdetails.User(user.getEmail(), user.getPassword(), new ArrayList<>());
+        User foundUser = userRepository.findByEmail(username).orElseThrow(() -> new UsernameNotFoundException(username));
+
+        return new UserDetailsImpl(foundUser.getId(), foundUser.getUserName(), foundUser.getPassword());
     }
 }
